added services folder and updated the folder structure
